refactor: make schema own module
