+++ src/lib.rs
@@ -7,4 +7,5 @@ extern crate regex;
 pub mod driver;
 pub mod error;
 pub mod query;
+pub mod schema;
 pub mod value;
